feat: restructure API project by creating user URLs, views, and serializers; update settings for ASGI and WSGI
