+++ Backend/API/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for APIapp project.
+Django settings for API project.
 
 Generated by 'django-admin startproject' using Django 5.1.4.
 
@@ -53,7 +53,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'APIapp.urls'
+ROOT_URLCONF = 'API.urls'
 
 TEMPLATES = [
     {
@@ -71,7 +71,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'APIapp.wsgi.application'
+WSGI_APPLICATION = 'API.wsgi.application'
 
 
 # Database
