+++ Backend/users/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+	path('', views.getUsersData),
+	path('create/', views.createUser),
+]
